+++ tests/helper.py
@@ -1,3 +1,4 @@
+import os
 from typing import Any
 
 from _pytest.pytester import RunResult
@@ -7,14 +8,18 @@ from pytest_mock import MockFixture
 
 def run_pytest(pytester: Pytester, mocker: MockFixture, *args: Any) -> RunResult:
     mocker.patch("pytest_helm_charts.fixtures.ExistingCluster", autospec=True)
+    mocker.patch.dict(
+        os.environ,
+        {
+            "KUBECONFIG": "/tmp/kat_test/kube.config",  # nosec: this is not used, mock value only
+            "ATS_CLUSTER_TYPE": "existing",
+            "ATS_EXTRA_EXTERNAL_CLUSTER_TYPE": "kind",
+        },
+        clear=True,
+    )
     result = pytester.runpytest(
-        "--cluster-type",
-        "existing",
         "--log-cli-level",
         "debug",
-        "--kube-config",
-        "/tmp/kat_test/kube.config",  # nosec
-        '--chart-extra-info="key1=val1,external_cluster_type=kind"',
         "-v",
         *args,
     )