+++ tests/test_plugin.py
@@ -1,14 +1,19 @@
-# noinspection PyProtectedMember
-from pytest_helm_charts.fixtures import _parse_extra_info  # noqa: F401
+from typing import Dict, Mapping
+
+import pytest
+
+from pytest_helm_charts.fixtures import _filter_extra_info_from_mapping, ENV_VAR_ATS_EXTRA_PREFIX
 
 
-def test_parse_chart_extra_info() -> None:
-    res = _parse_extra_info("key1=val1,external_cluster_type=kind")
-    assert len(res) == 2
-    assert "key1" in res and res["key1"] == "val1"
-    assert "external_cluster_type" in res and res["external_cluster_type"] == "kind"
-
-
-def test_parse_chart_no_extra_info() -> None:
-    res = _parse_extra_info("")
-    assert len(res) == 0
+@pytest.mark.parametrize(
+    ["env", "expected"],
+    [
+        ({}, {}),
+        ({"t": "a"}, {}),
+        ({ENV_VAR_ATS_EXTRA_PREFIX + "TEST": "abc"}, {"test": "abc"}),
+        ({ENV_VAR_ATS_EXTRA_PREFIX + "CHUCK": "Norris", "Bruce": "Lee"}, {"chuck": "Norris"}),
+    ],
+    ids=["empty", "none matching", "single and matching", "some matching"],
+)
+def test_test_extra_info(env: Mapping[str, str], expected: Dict[str, str]) -> None:
+    assert _filter_extra_info_from_mapping(env) == expected